worker/requirements.txt
redis==6.4.0
pymongo==4.15.3
python-dotenv==1.1.1

